backend/itegrateSys/requirements.txt
asgiref==3.3.1
backcall==0.2.0
cffi==1.14.4
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
djangorestframework==3.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jwt==1.1.0
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
PyMySQL==0.10.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
pycryptodomex
pycryptodome